+++ tests/test_external_models.py
@@ -122,6 +122,20 @@ def test_IttiKoch(tmpdir, matlab, color_stimulus, grayscale_stimulus):
 
 @pytest.mark.skip_octave
 @pytest.mark.matlab
+def test_RARE2007(tmpdir, matlab, color_stimulus, grayscale_stimulus):
+    model = pysaliency.external_models.RARE2007(location=str(tmpdir))
+    print('Testing color')
+    saliency_map = model.saliency_map(color_stimulus)
+    np.testing.assert_allclose(saliency_map,
+                               np.load(os.path.join('tests', 'external_models', '{}_color_stimulus.npy'.format(model.__modelname__))))
+    print('Testing Grayscale')
+    saliency_map = model.saliency_map(grayscale_stimulus)
+    np.testing.assert_allclose(saliency_map,
+                               np.load(os.path.join('tests', 'external_models', '{}_grayscale_stimulus.npy'.format(model.__modelname__))))
+
+
+@pytest.mark.skip_octave
+@pytest.mark.matlab
 def test_RARE2012(tmpdir, matlab, color_stimulus, grayscale_stimulus):
     model = pysaliency.RARE2012(location=str(tmpdir))
     print('Testing color')